Move internal/merger to merger
