+++ cmd/gazelle/BUILD.bazel
@@ -17,12 +17,12 @@ go_library(
     deps = [
         "//config:go_default_library",
         "//flag:go_default_library",
-        "//internal/merger:go_default_library",
         "//internal/version:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
         "//language/go:go_default_library",
         "//language/proto:go_default_library",
+        "//merger:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
