+++ language/go/BUILD.bazel
@@ -73,10 +73,10 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//config:go_default_library",
-        "//internal/merger:go_default_library",
         "//label:go_default_library",
         "//language:go_default_library",
         "//language/proto:go_default_library",
+        "//merger:go_default_library",
         "//repos:go_default_library",
         "//resolve:go_default_library",
         "//rule:go_default_library",
